Adicionando configuracao do S3 Storage

+++ src/controllers/PostController.js
@@ -12,19 +12,25 @@ module.exports = {
     async store(req, res) {
 
         const { author, place, description, hashtags } = req.body;
-        const { filename: image } = req.file;  
+        const { filename: image, key, location: url = "" } = req.file;  
         
-        const [name] = image.split('.');
-        const fileName = `${name}.jpg`;
+        let fileName = url;
 
-        await sharp(req.file.path)
-        .resize(500)
-        .jpeg({ quality: 70 })
-        .toFile(
-            path.resolve(req.file.destination, 'resized', fileName) 
-        );
+        if (process.env.STORAGE_TYPE === 'local'){
 
-        fs.unlinkSync(req.file.path);
+            const [name] = image.split('.');
+            fileName = `${name}.jpg`;
+    
+            await sharp(req.file.path)
+            .resize(500)
+            .jpeg({ quality: 70 })
+            .toFile(
+                path.resolve(req.file.destination, 'resized', fileName) 
+            );
+    
+            fs.unlinkSync(req.file.path);
+
+        }
 
         const post = await Post.create({
             author,
